Add index.html & recipes folder with recipe in it
Created index.html which will be the main page of the website. This web
page has just a heading and a link to the one recipe for now.

Also created a recipes subfolder in the directory that will house all
the recipe web pages I will create later.

Inside this recipes subfolder, there is a baconchicken.html folder that
will have the content for the first recipe. Right now, it has a simple
heading and a picture of the finished recipe.